frontend/requirements.txt

streamlit>=1.33
requests
python-dotenv
streamlit-lottie
Pillow